front: test e2e - add rolling stock selector page
